[#37] Refactored folder structure
