--- src/common/borderBottom.js
@@ -1,8 +1,0 @@
-import React from 'react';
-import styles from './borderBottom.module.scss'; 
-
-const BorderBottom = () => {
-    return <div className={styles.borderBottom}/>
-}
-
-export default BorderBottom;
