--- src/common/borderBottom.module.scss
@@ -1,3 +1,0 @@
-@import 'src/scss/variables.scss';
-
-.borderBottom {border-bottom: 2px solid $grey1};
