+++ src/components/_shared/line.js
@@ -1,0 +1,8 @@
+import React from 'react';
+import styles from './line.module.scss'; 
+
+const Line = () => {
+    return <div className={styles.line}/>
+}
+
+export default Line;
