+++ src/components/_shared/line.module.scss
@@ -1,0 +1,3 @@
+@import 'src/scss/variables.scss';
+
+.line {border-bottom: 2px solid $grey1};
