+++ src/components/players/details/games.module.scss
