+++ src/components/players/details/index.module.scss
@@ -20,7 +20,7 @@
       }
     }
 
-    .imageBorderBottom {
+    .imageLine {
       width: 100%;
       margin-top: 5px;
       border-bottom: 3px solid $background-color-purple;
