+++ src/components/players/details/stats/index.js
