+++ src/components/players/details/stats/placeholder.module.scss
