+++ src/components/players/list/index.module.scss
@@ -28,7 +28,7 @@
   .imageContainer {
     display: flex;
     padding-top: 10px;
-    background-image: url('../../assets/imgs/player-backdrop.png');
+    background-image: url('../../../assets/imgs/player-backdrop.png');
     background-size: cover;
 
     img {
@@ -36,7 +36,7 @@
     }
   }
 
-  .imageBorderBottom {
+  .imageLine {
     width: 100%;
     margin-top: 4px;
     border-bottom: 2px solid $background-color-purple;
