+++ src/jsConfig.json
@@ -1,0 +1,6 @@
+{
+"compilerOptions": {
+    "baseUrl": "src", 
+    "include": ["src"]
+  }
+}
