+++ src/scss/variables.scss
@@ -6,7 +6,7 @@ $border-shadow: #D4D3D3;
 $border-dark-grey: #919191;
 $grey-three: #919191;
 $white: #FFFFFF;
-$greyBorderBottom:#D8D8D8;
+$greyLine:#D8D8D8;
 $won: rgb(53, 211, 53);
 $lost:  rgb(233, 30, 30);
